PageFooter 컴포넌트 파일로 분리, 책공유 페이지에 삽입
