+++ client/src/BookSharingPage.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import VerticalContainer from './common/VerticalContainer';
-import PageHeader from './PageHeader';
+import PageHeader from './common/PageHeader';
 import SearchBox from './common/SearchBox';
+import PageFooter from './common/PageFooter';
 
 function BookSharingPage() {
   return (
@@ -13,6 +14,7 @@ function BookSharingPage() {
         subPhrase="당신도 다른 사람들과, 책을 통해서 이야기를 만들어 나갈 수 있어요.
         주소를 알려주시면 주변의 책을 알려 드릴게요."
       />
+      <PageFooter />
     </VerticalContainer>
   );
 }
