+++ client/src/MainPage.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import styled from 'styled-components';
-import PageHeader from './PageHeader';
+import PageHeader from './common/PageHeader';
 import book1 from './image/book1.jpg';
 import book2 from './image/book2.jpg';
 import book3 from './image/book3.jpg';
@@ -10,6 +9,7 @@ import SlideMenu from './common/SlideMenu';
 import SearchBox from './common/SearchBox';
 import VerticalContainer from './common/VerticalContainer';
 import SmallGroupSummary from './SmallGroupSummary';
+import PageFooter from './common/PageFooter';
 
 const slideMenuImageList = [book1, book2, book3, book4, book5];
 
@@ -19,67 +19,6 @@ const smallGroupTitleList = [
   '소모임 모집글 3',
   '소모임 모집글 4',
 ];
-
-const siteMenuList = [
-  '자유게시판',
-  '리뷰게시판',
-  '어떤게시판',
-  '게시판이름',
-];
-
-const PageFooterContainer = styled.div`
-  width:100%;
-  background: #343a40;
-  color:white;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  margin-top:1rem;
-  padding-bottom: 5rem;
-`;
-
-const SiteMapListContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: flex-start;
-  margin:0 0.3rem;
-`;
-
-const SiteMapListTitle = styled.h2`
-  font-size: 1.5rem;
-  text-align: left;
-`;
-
-const SiteMapListItem = styled.h3`
-  margin:0;
-  font-size: 1rem;
-  font-weight: normal;
-`;
-
-function SiteMapList({ title, menuList }) {
-  return (
-    <SiteMapListContainer>
-      <SiteMapListTitle>{title}</SiteMapListTitle>
-      {menuList.map((item) => <SiteMapListItem key={item}>{item}</SiteMapListItem>)}
-    </SiteMapListContainer>
-  );
-}
-
-function PageFooter() {
-  return (
-    <PageFooterContainer>
-      <SiteMapListContainer>
-        <SiteMapList title="게시판1" menuList={siteMenuList} />
-      </SiteMapListContainer>
-      <SiteMapListContainer>
-        <SiteMapList title="게시판2" menuList={siteMenuList} />
-      </SiteMapListContainer>
-      <SiteMapListContainer>
-        <SiteMapList title="게시판3" menuList={siteMenuList} />
-      </SiteMapListContainer>
-    </PageFooterContainer>
-  );
-}
 
 function MainPage() {
   return (
